simulator/images/requirements.txt
Pillow>=10
yamcs-client>=1.9